move lattice lstm out of specialized encoders package
